refactor: ResponseSender 클래스를 별도의 파일로 이동
