+++ src/server/Server.hpp
@@ -16,31 +16,9 @@
 #include "../http/model/Packet.hpp"
 #include "../http/serializer/Serializer.hpp"
 #include "epoll/manager/EpollManager.hpp"
+#include "sender/ResponseSender.hpp"
 
 namespace server {
-	enum class SendResult {
-		SUCCESS,
-		RETRY,
-		ERROR
-	};
-
-	class ResponseSender {
-		private:
-			std::string _responseBuffer;
-			int _socketFd;
-			size_t _responseSent;
-			ResponseSender();
-
-		public:
-			ResponseSender(int socketFd, const http::Packet& httpPacket) :
-				_responseBuffer(http::Serializer::serialize(httpPacket)),
-				_socketFd(socketFd),
-				_responseSent(0) {}
-			ResponseSender(int socketFd, const std::string& rawResponse) :
-				_responseBuffer(rawResponse), _socketFd(socketFd), _responseSent(0) {}
-			SendResult send();
-	};
-
 	class Server {
 		private:
 			std::map<int, config::Config> _configs;
@@ -59,8 +37,6 @@ namespace server {
 			void initServer(int);
 			void loop();
 			void handleEvents();
-			// bool sendResponse(int, const http::Packet&);
-			// bool sendResponse(int, const std::string&);
 
 		public:
 			explicit Server(const std::map<int, config::Config>&);
